What the walk asks with is beside it rather than inside it
The check was one file because everything it needed was written where it was
needed. What it needs is not one thing, though, and each of them is a question
with a single answer that several readers want — which is the shape a file
should have, and is what the walk was hiding.

So the walk keeps what only it can say: where to look next, what is known where
it is, what a construction comes out as, and what to report. Beside it now:

  DischargeRules — what the language's own operations do to the properties this
      check tracks. Data and lookups, keyed by the operation a call reaches. What
      a rule is used for is not here, because that is the walk's.
  Clauses — a declaration's invariant, typed once over the fields it is written
      against and read at a value by putting what each field is given where that
      field is read. It remembers, because the seeding and every construction ask
      the same declaration the same questions.
  Terms — where a value is, what it is called, and whether anything can be said
      of it. Three answers, and they are not the same question: a location is
      what the seeding writes about, a key is how two expressions are known to
      compute one value, and being readable is whether a clause read against a
      value could ever be discharged.
  Predicates — what a clause owes where a value is built, what a guard settles
      where one is written, and the keys by which the two meet. One language read
      in two directions, and written apart the two would drift — every drift a
      clause an author guarded and the check still reported.
  Known, Quantified, Denotations, Denotes — the values threaded through all of
      it, which were nested records in the one file and are what any of these
      would have had to import from the walk.

Nothing here changes what the check answers. The whole suite is green — 2709 in
the compiler, 151 in the language server, 109 in the syntax, 1 in the CLI — and
warm compile is 138.7 ms for crm and 18.1 ms for issuetracker, against 144.8 and
18.1 before any of this.

2397 lines in one file become 685 for the walk and 1723 in five files it reads
with.
